pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
description = "LiveKit voice agent with Speechmatics integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "speechmatics-rt>=1.0.0",
    "speechmatics-voice[smart]>=0.2.8",
    "livekit-agents==1.4.2",
    "livekit-plugins-speechmatics==1.4.2",
    "livekit-plugins-elevenlabs==1.4.2",
    "livekit-plugins-openai==1.4.2",
    "livekit-plugins-deepgram==1.4.2",
    "livekit-plugins-soniox==1.4.2",
    "livekit-plugins-silero==1.4.2",
    "livekit-plugins-turn-detector==1.4.2",
    "python-dotenv>=1.2.1",
    "pydub>=0.25",
    "sounddevice>=0.5",
    "numpy>=2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.9",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I", "F"]
